add amount to view in customer

+++ wsh/wsh-dvd-rental/src/main/java/com/wsh/controller/payment/PaymentAddController.java
@@ -8,13 +8,15 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.math.BigDecimal;
+
 @Controller
 @RequiredArgsConstructor
 @RequestMapping("payment")
 public class PaymentAddController {
     private final PaymentService paymentService;
     @PostMapping("/add")
-    public String add( @RequestParam("paymentAmount") final Double amount){
+    public String add( @RequestParam("paymentAmount") final BigDecimal amount){
         paymentService.add(amount);
         return "redirect:/payment/list";
     }
