+++ wsh/wsh-dvd-rental/src/main/java/com/wsh/controller/payment/PaymentEditController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.beans.Transient;
+import java.math.BigDecimal;
 
 @Controller
 @RequiredArgsConstructor
@@ -16,7 +17,7 @@ public class PaymentEditController {
     private final PaymentService paymentService;
 
     @PostMapping("/update")
-    public String update(final @RequestParam("paymentId") Long paymentID,final @RequestParam("paymentAmount") Double amount){
+    public String update(final @RequestParam("paymentId") Long paymentID,final @RequestParam("paymentAmount") BigDecimal amount){
         paymentService.update(paymentID,amount);
         return "redirect:/payment/list";
     }
