+++ wsh/wsh-dvd-rental/src/main/java/com/wsh/domain/Payment.java
@@ -3,6 +3,7 @@ package com.wsh.domain;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -13,14 +14,14 @@ public class Payment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long paymentId;
-    private Double amount;
+    private BigDecimal amount;
     private LocalDateTime paymentDate = LocalDateTime.now();
     private LocalDateTime lastUpdate = LocalDateTime.now();
     @ManyToOne(optional = false, fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     @JoinColumn(name = "customer_id")
     private Customer customer;
 
-    public Payment(Double amount) {
+    public Payment(BigDecimal amount) {
         this.amount = amount;
     }
 }
